requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
redis==5.0.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
aiohttp==3.9.3
